Fix 500 errors on file upload
There were errors during file upload due to a missing S3 bucket name
parameter. This ensures that is added by setting permissions between the
function(s) and the bucket.
