--- infrastructure/lib/queue-permissions.ts
@@ -1,7 +1,0 @@
-/**
- * Supported Queue Permissions for the application.
- */
-export enum QueuePermissions {
-  SEND = "SEND",
-  CONSUME = "CONSUME",
-}
