--- infrastructure/lib/table-permissions.ts
@@ -1,9 +1,0 @@
-/**
- * Supported Table Permissions for the application.
- */
-export enum TablePermissions {
-  ALL = "ALL",
-  READ = "READ",
-  READ_WRITE = "READ_WRITE",
-  WRITE = "WRITE",
-}
